French translation for module "Graph" (#634)
The first part of the translation is a direct translation of the
english part. The second is too, although it wasn't done by me
but by wikipedia contributors (the second paragraph of the
English wikipedia introduction is exactly the same as the first
paragraph of the French wikipedia "Description")

Add french version redirection on the README.md. The french
version is placed before the portuguese version to have a
cleaner diff (no trailing commas to edit)